chore(tinlake-ui): move Unlock NFT button
